use TextEditor component in editor modal
