+++ apps/studio/src/components/tableview/EditorModal.vue
@@ -41,7 +41,11 @@
         </div>
 
         <div class="editor-container">
-          <textarea name="editor" ref="editorRef" />
+          <text-editor
+            ref="textEditor"
+            :lang="languageName"
+            :initializeOnMount="false"
+          />
         </div>
       </div>
       <div class="bottom">
@@ -90,7 +94,6 @@
 
 <script lang="ts">
 import Vue from 'vue'
-import CodeMirror from 'codemirror'
 import 'codemirror/addon/comment/comment'
 import 'codemirror/keymap/vim.js'
 import 'codemirror/addon/dialog/dialog'
@@ -101,11 +104,11 @@ import 'codemirror/addon/search/matchesonscrollbar'
 import 'codemirror/addon/search/matchesonscrollbar.css'
 import 'codemirror/addon/search/searchcursor'
 import { Languages, LanguageData, TextLanguage, getLanguageByName, getLanguageByContent } from '../../lib/editor/languageData'
-import setKeybindingsFromVimrc from '@/lib/readVimrc'
 import { uuidv4 } from "@/lib/uuid"
 import _ from 'lodash'
 import { mapGetters } from 'vuex'
 import rawlog from 'electron-log'
+import TextEditor from '@/components/common/TextEditor.vue'
 
 const log = rawlog.scope('EditorModal')
 
@@ -122,7 +125,7 @@ export default Vue.extend({
       wrapText: TextLanguage.wrapTextByDefault,
     }
   },
-
+  components: { TextEditor },
   computed: {
     ...mapGetters({ 'settings': 'settings/settings' }),
     modalName() {
@@ -187,56 +190,15 @@ export default Vue.extend({
     async onOpen() {
       await this.$nextTick();
 
-      const language = this.language
-
-      this.editor = CodeMirror.fromTextArea(this.$refs.editorRef, {
-        lineNumbers: true,
-        lineWrapping: this.wrapText,
-        mode: language !== null ? language.editorMode : undefined,
-        indentWithTabs: false,
-        tabSize: 2,
-        theme: 'monokai',
-        extraKeys: {
-          "Ctrl-Space": "autocomplete",
-          "Shift-Tab": "indentLess",
-          [this.cmCtrlOrCmd('F')]: 'findPersistent',
-          [this.cmCtrlOrCmd('R')]: 'replace',
-          [this.cmCtrlOrCmd('Shift-R')]: 'replaceAll'
-        },
-        options: {
-          closeOnBlur: false
-        },
-        keyMap: this.userKeymap
-      } as any)
-
+      this.editor = this.$refs.textEditor.initialize()
+      this.editor.setOption("lineWrapping", this.wrapText)
       this.editor.setValue(this.content)
-      this.editor.on("keydown", (_cm, e) => {
-        if (this.$store.state.menuActive) {
-          e.preventDefault()
-        }
-      })
-
-      if (this.userKeymap === "vim") {
-        const codeMirrorVimInstance = document.querySelector(".CodeMirror").CodeMirror.constructor.Vim
-        if (!codeMirrorVimInstance) {
-          log.error("Could not find code mirror vim instance");
-        } else {
-          setKeybindingsFromVimrc(codeMirrorVimInstance);
-        }
-      }
-
       this.editor.on("change", (cm) => {
         this.content = cm.getValue()
         this.debouncedCheckForErrors();
       })
 
-      this.editor.focus()
-
-      setTimeout(() => {
-        // this fixes the editor not showing because it doesn't think it's dom element is in view.
-        // its a hit and miss error
-        this.editor.refresh()
-      }, 1)
+      this.$refs.textEditor.focus()
 
       this.$nextTick(this.resizeHeightToFitContent)
     },
