+++ apps/studio/src/components/tableview/TableTable.vue
@@ -634,7 +634,7 @@ export default Vue.extend({
       return this.tablesInitialLoaded && this.active && !this.initialized
     },
     columnFilterModalName() {
-      return `column-filter-modal-${this.tab.id}`
+      return `column-filter-modal-${this.tableId}`
     },
   },
